--- a/Project.toml
+++ b/Project.toml
@@ -9,7 +9,7 @@ MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 
 [compat]
-MLStyle = "0.3"
+MLStyle = "0.4"
 TupleTools = "1.2"
 DataStructures = "0.17"
 julia = "1.3,1.4"