@thautwarm 2020-05-10
Hi, it seems there's no significant change, so I'd say **NOT** to apply this change. I'll try to work on the performance improvements.