@GiggleLiu 2020-05-10
Thanks, but no worry, I think NiLang is already very fast! The bottle neck is not nessesarily MLStyle.

Have a rest!